+++ shared/src/main/scala/scalus/builtin/ToDataMacros.scala
@@ -1,77 +1,17 @@
-package scalus.prelude
+package scalus.builtin
 
 import scalus.CompileDerivations
-import scalus.builtin.Builtins
 
+import scala.collection.immutable.List
 import scala.quoted.*
 
-type Data = scalus.builtin.Data
-type ByteString = scalus.builtin.ByteString
-
-@FunctionalInterface
-trait ToData[-A] extends Function1[A, Data] with CompileDerivations {
-    override def apply(v1: A): Data
-}
-
-@scalus.Compile
-object ToData {
-
-    inline def derived[A]: ToData[A] = ${
-        DataRepresentation.toDataImpl[A]
-    }
-
-    given scalus.prelude.ToData[Data] = (a: Data) => a
-    given scalus.prelude.ToData[ByteString] = Builtins.bData
-    given scalus.prelude.ToData[BigInt] = Builtins.iData
-
-}
-
-@FunctionalInterface
-trait FromData[+A] extends Function1[Data, A] with CompileDerivations {
-    override def apply(v: Data): A
-}
-
-@scalus.Compile
-object FromData {
-
-    inline def derived[A]: FromData[A] = ${
-        DataRepresentation.fromDataImpl[A]
-    }
-
-    given FromData[Data] = (a: Data) => a
-    given FromData[ByteString] = Builtins.unBData
-    given FromData[BigInt] = Builtins.unIData
-    given FromData[String] = (d: Data) => Builtins.decodeUtf8(Builtins.unBData(d))
-
-}
-
-trait DataRepresentation[A] {
-
-    def toData(a: A): Data
-
-    def fromData(d: Data): A
-
-    def select(d: Data, fieldName: String): Data
-
-}
-
-object DataRepresentation {
+object ToDataMacros {
 
     def toDataImpl[A: Type](using Quotes): Expr[ToData[A]] = {
         import quotes.reflect.*
         '{ (a: A) =>
             ${ generateToDataApply[A]('a) }
         }
-        /*
-        '{
-            new ToData[A] {
-                override def apply(a: A): Data = {
-                    ${ generateToDataApply[A]('a) }
-                }
-            }
-        }
-
-         */
     }
 
     def generateToDataApply[A: Type](a: Expr[A])(using Quotes): Expr[Data] = {
@@ -147,7 +87,7 @@ object DataRepresentation {
                       .map { case (binding, tpe) =>
                           tpe.widen.asType match
                               case '[t] =>
-                                  Expr.summon[scalus.prelude.ToData[t]] match
+                                  Expr.summon[ToData[t]] match
                                       case None =>
                                           report.errorAndAbort(
                                             s"Could not find given ToData[${tpe.widen.show}] within ${TypeRepr.of[A].show} wheb generate product"
@@ -185,7 +125,7 @@ object DataRepresentation {
                       .map { case (binding, atpe) =>
                           atpe.asType match
                               case '[t] =>
-                                  Expr.summon[scalus.prelude.ToData[t]] match
+                                  Expr.summon[ToData[t]] match
                                       case None =>
                                           println(
                                             s"atpe=${atpe.show}, atpe.widen=${atpe.widen.show}, atpe."
@@ -253,130 +193,145 @@ object DataRepresentation {
         genMatch(value.asTerm).asExprOf[Data]
     }
 
-    def fromDataImpl[A: Type](using Quotes): Expr[FromData[A]] = {
+    def deriveEnumMacro[T: Type](using Quotes): Expr[ToData[T]] = {
         import quotes.reflect.*
-        val ta = TypeRepr.of[A].dealias.widen
-        if ta <:< TypeRepr.of[AnyRef] then
-            val children = ta.typeSymbol.children
-            if children.isEmpty then
-                if ta.typeSymbol.flags.is(Flags.Trait) then
-                    report.errorAndAbort(
-                      s"Cannot derive FromData for trait ${ta.typeSymbol.fullName}"
-                    )
-                else if ta.typeSymbol.flags.is(Flags.Case | Flags.Enum) then {
-                    deriveFromDataCaseClassApply[A]
-                } else {
-                    report.errorAndAbort(
-                      s"Cannot derive FromData for ${ta.typeSymbol.fullName} which is not a case class or enum"
-                    )
-                }
-            else {
-                deriveFromDataSumCaseClassApply[A]
-            }
-        else
-            report.errorAndAbort(
-              s"Cannot derive FromData for ${ta.typeSymbol.fullName} which is not a case class or enum"
-            )
-    }
 
-    def deriveFromDataCaseClassApply[A: Type](using Quotes): Expr[FromData[A]] = {
-        '{ (d: Data) =>
-            val args = scalus.builtin.Builtins.unConstrData(d).snd
-
-            // generate f = (args) => new Constructor(args.head, args.tail.head, ...)
-            // then apply to args: f(args)
-            // and finally beta reduce it in compile time
-            ${ Expr.betaReduce('{ ${ deriveFromDataCaseClassConstructor[A] }(args) }) }
-        }
-    }
-
-    private[scalus] def deriveFromDataCaseClassConstructor[T: Type](using
-        Quotes
-    ): Expr[scalus.builtin.List[Data] => T] = {
-        import quotes.reflect.*
-        val classSym = TypeTree.of[T].symbol
-        val constr = classSym.primaryConstructor
-        val params = constr.paramSymss.flatten
-        val fromDataOfArgs = params.map { param =>
-            val tpe = param.termRef.widen.dealias
-            tpe.asType match
-                case '[t] =>
-                    Expr.summon[FromData[t]] match
-                        case None =>
-                            report.errorAndAbort(
-                              s"Could not find given FromData[${tpe.show}] in case class constructor generation for ${TypeRepr.of[T].show}]"
-                            )
-                        case Some(value) => value
-        }
-
-        def genGetter(init: Expr[scalus.builtin.List[Data]], idx: Int)(using Quotes): Expr[Data] =
-            var expr = init
-            var i = 0
-            while i < idx do
-                val exp = expr // save the current expr, otherwise it will loop forever
-                expr = '{ $exp.tail }
-                i += 1
-            '{ $expr.head }
-
-        def genConstructorCall(
-            a: Expr[scalus.builtin.List[scalus.builtin.Data]]
-        )(using Quotes): Expr[T] = {
-            val args = fromDataOfArgs.zipWithIndex.map { case (appl, idx) =>
-                val arg = genGetter(a, idx)
-                '{ $appl($arg) }.asTerm
-            }
-            // Couldn't find a way to do this using quotes, so just construct the tree manually
-            New(TypeTree.of[T]).select(constr).appliedToArgs(args).asExprOf[T]
-        }
-
-        '{ (args: scalus.builtin.List[scalus.builtin.Data]) => ${ genConstructorCall('{ args }) } }
-
-    }
-
-    def deriveFromDataSumCaseClassApply[A: Type](using Quotes): Expr[FromData[A]] = {
-        import quotes.reflect.*
-        val constrTpe = TypeRepr.of[A]
-        val typeSymbol = TypeRepr.of[A].widen.dealias.typeSymbol
+        val constrTpe = TypeRepr.of[T]
+        val typeSymbol = TypeRepr.of[T].widen.dealias.typeSymbol
         if !typeSymbol.flags.is(Flags.Enum) then
             report.errorAndAbort(
               s"deriveEnum can only be used with enums, got ${typeSymbol.fullName}"
             )
 
-        val mappingRhs: scala.List[Expr[scalus.builtin.List[Data] => A]] = typeSymbol.children.map {
-            child =>
-                child.typeRef.asType match
-                    case '[t] =>
-                        // println(s"child: ${child}, ${child.flags.show} ${child.caseFields}")
-                        if child.caseFields.isEmpty then
-                            '{ (_: scalus.builtin.List[Data]) =>
-                                ${ Ident(child.termRef).asExprOf[t] }
-                            }.asExprOf[scalus.builtin.List[Data] => A]
-                        else {
-                            deriveFromDataCaseClassConstructor[t]
-                                .asExprOf[scalus.builtin.List[Data] => A]
-                        }
-                    case _ =>
-                        report.errorAndAbort(
-                          s"Cannot derive FromData for ${child.typeRef.show} "
-                        )
-
+        def genRhs(constrIdx: Int, bindings: List[(Symbol, TypeRepr)])(using Quotes) = '{
+            Builtins.constrData(
+              BigInt(${ Expr(constrIdx) }),
+              ${
+                  val args = bindings
+                      .map { case (binding, tpe) =>
+                          tpe.asType match
+                              case '[t] =>
+                                  Expr.summon[ToData[t]] match
+                                      case None =>
+                                          report.errorAndAbort(
+                                            s"Could not find given ToData[${tpe.widen.show}]"
+                                          )
+                                      case Some(toData) =>
+                                          val arg = Ident(binding.termRef).asExprOf[t]
+                                          '{ $toData($arg) }
+                      }
+                      .asInstanceOf[List[Expr[Data]]]
+                  args.foldRight('{ scalus.builtin.Builtins.mkNilData() }) { (data, acc) =>
+                      '{ scalus.builtin.Builtins.mkCons($data, $acc) }
+                  }
+              }
+            )
         }
-        // .asInstanceOf[scala.List[(Expr[scalus.builtin.List[Data] => A], Int)]]
 
-        // stage programming is cool, but it's hard to comprehend what's going on
-        '{ (d: Data) =>
-            val pair = Builtins.unConstrData(d)
-            val tag = pair.fst
-            val args = pair.snd
-            ${
-                mappingRhs.zipWithIndex.foldRight('{
-                    throw new Exception("Invalid tag")
-                }.asExprOf[A]) { case ((code, t), acc) =>
-                    '{
-                        if Builtins.equalsInteger(tag, BigInt(${ Expr(t) })) then $code(args)
-                        else $acc
+        def genMatch(prodTerm: Term)(using Quotes) = {
+            val cases = typeSymbol.children.zipWithIndex.map { (childTypeSymbol, tag) =>
+                if childTypeSymbol.caseFields.isEmpty then
+                    val rhs = genRhs(tag, Nil).asTerm
+                    CaseDef(Ident(childTypeSymbol.termRef), None, rhs)
+                else
+                    val classSym = childTypeSymbol
+                    val companionModuleRef = classSym.companionModule
+                    val unapplyRef = companionModuleRef.methodMember("unapply").head.termRef
+                    val constr = classSym.primaryConstructor
+                    val params = constr.paramSymss.flatten
+                    val paramsNameType = params.map(p => p.name -> p.typeRef)
+                    val bindingsSymbols = paramsNameType.map { (name, tpe) =>
+                        (Symbol.newBind(Symbol.noSymbol, name, Flags.EmptyFlags, tpe), tpe)
                     }
-                }
+
+                    val bindings = bindingsSymbols.map { (symbol, _) =>
+                        Bind(symbol, Wildcard())
+                    }
+                    val rhs = genRhs(tag, bindingsSymbols).asTerm
+                    CaseDef(
+                      Typed(
+                        Unapply(Ident(unapplyRef), Nil, bindings).asInstanceOf[Term],
+                        TypeTree.ref(childTypeSymbol)
+                      ),
+                      None,
+                      rhs
+                    )
+            }
+            val m =
+                Match(prodTerm, cases)
+            m
+        }
+
+        '{ (product: T) =>
+            ${
+                val prodTerm = '{ product }.asTerm
+                val code = genMatch(prodTerm).asExprOf[Data]
+                code
+            }
+        }
+    }
+
+    def deriveCaseClassMacro[T: Type](constrIdx: Expr[Int])(using Quotes): Expr[ToData[T]] = {
+        import quotes.reflect.*
+        val classSym = TypeTree.of[T].symbol
+        val companionModuleRef = classSym.companionModule
+        val unapplyRef = companionModuleRef.methodMember("unapply").head.termRef
+        val constr = classSym.primaryConstructor
+        val params = constr.paramSymss.flatten
+        val paramsNameType = params.map(p => p.name -> p.typeRef)
+
+        /*
+      Generate a pattern match to introduce all the params,
+      to avoid a.field1, a.field2, etc.
+      Something ike:
+        a match
+          case A(field1, field2, ...) =>
+            mkConstr(
+              BigInt($constrIdx),
+              mkCons(field1.toData, mkCons(field2.toData, ...))
+            )
+         */
+        def genMatch(prodTerm: Term, params: List[(String, TypeRepr)])(using Quotes) = {
+            val bindingsSymbols = params.map { (name, tpe) =>
+                (Symbol.newBind(Symbol.noSymbol, name, Flags.EmptyFlags, tpe), tpe)
+            }
+
+            val bindings = bindingsSymbols.map { (symbol, _) =>
+                Bind(symbol, Wildcard())
+            }
+            val rhs = genRhs(bindingsSymbols).asTerm
+            Match(prodTerm, List(CaseDef(Unapply(Ident(unapplyRef), Nil, bindings), None, rhs)))
+        }
+
+        def genRhs(bindings: List[(Symbol, TypeRepr)])(using Quotes) = '{
+            Builtins.constrData(
+              BigInt($constrIdx),
+              ${
+                  val args = bindings
+                      .map { case (binding, tpe) =>
+                          tpe.asType match
+                              case '[t] =>
+                                  Expr.summon[ToData[t]] match
+                                      case None =>
+                                          report.errorAndAbort(
+                                            s"Could not find given ToData[${tpe.widen.show}]"
+                                          )
+                                      case Some(toData) =>
+                                          val arg = Ident(binding.termRef).asExprOf[t]
+                                          '{ $toData($arg) }
+                      }
+                      .asInstanceOf[List[Expr[Data]]]
+                  args.foldRight('{ scalus.builtin.Builtins.mkNilData() }) { (data, acc) =>
+                      '{ scalus.builtin.Builtins.mkCons($data, $acc) }
+                  }
+              }
+            )
+        }
+
+        '{ (product: T) =>
+            ${
+                val prodTerm = '{ product }.asTerm
+                genMatch(prodTerm, paramsNameType).asExprOf[Data]
             }
         }
