+++ shared/src/test/scala/scalus/prelude/DerivingSpec.scala
@@ -9,18 +9,46 @@ import scalus.uplc.eval.{*, given}
 //import scalus.builtin.given
 import scalus.Compiler.{compile, compileDebug}
 import scalus.testutil.SIRModules
+import scalus.builtin.*
+import scalus.builtin.ToDataInstances.given
+import scalus.builtin.FromDataInstances.given
+
+/*
+
+TODO: enble after
+  - compiler with fix of https://github.com/scala/scala3/issues/23043
+  - implementing of automatic generation of 'derived' methods on SIR level.
 
 enum DerivingSpec_AE2 derives ToData, FromData:
     case DS2_A extends DerivingSpec_AE2
     case DS2_B(b: BigInt) extends DerivingSpec_AE2
     case DS2_C(b: BigInt, bs: ByteString) extends DerivingSpec_AE2
 
+ */
+
 object DerivingSpecScope {
 
+    /*
     enum AE1 derives ToData, FromData:
         case A extends AE1
         case B(b: BigInt) extends AE1
         case C(b: BigInt, bs: ByteString) extends AE1
+
+     */
+
+}
+
+enum DerivingSpec_AE3 {
+    case DS3_A extends DerivingSpec_AE3
+    case DS3_B(b: BigInt) extends DerivingSpec_AE3
+    case DS3_C(b: BigInt, bs: ByteString) extends DerivingSpec_AE3
+}
+
+@Compile
+object DerivingSpec_AE3 {
+
+    given ToData[DerivingSpec_AE3] = ToData.derived
+    given FromData[DerivingSpec_AE3] = FromData.derived
 
 }
 
@@ -31,6 +59,7 @@ class DerivingSpec extends AnyFunSuite {
     import scalus.builtin.given
     protected given PlutusVM = PlutusVM.makePlutusV3VM()
 
+    /*
     test("Compile To/From Data for AE1") {
 
         val sir = compileDebug { (d: Data) =>
@@ -69,6 +98,43 @@ class DerivingSpec extends AnyFunSuite {
 
                 fail(s"Expected success, but got failure, logs=$logs")
 
+    }
+    
+     */
+
+    test("Compile To/From Data for AE3") {
+
+        val sir = Compiler.compileDebug { (d: Data) =>
+            val a = summon[FromData[DerivingSpec_AE3]](d)
+            a match
+                case DerivingSpec_AE3.DS3_A        => BigInt(1)
+                case DerivingSpec_AE3.DS3_B(b)     => BigInt(2)
+                case DerivingSpec_AE3.DS3_C(b, bs) => BigInt(3)
+        }
+
+        println(s"sir: ${sir.pretty.render(1000)}")
+
+        val uplc = sir.toUplc(generateErrorTraces = true)
+        val ae3 = DerivingSpec_AE3.DS3_A
+        val ae3Data = summon[ToData[DerivingSpec_AE3]](ae3)
+        val program1 = uplc.plutusV3 $ ae3Data
+        val result1 = program1.evalDebug
+        // assert(result1.isSuccess)
+        result1 match
+            case Result.Success(term, _, _, _) =>
+                assert(term == Term.Const(Constant.Integer(1)))
+            case Result.Failure(e, _, _, logs) =>
+                e.printStackTrace()
+                println(s"logs=${logs}")
+
+                println("loading module")
+                val module = SIRModules.load("scalus.prelude.DerivingSpec_AE3$")
+
+                module.defs.foreach { b =>
+                    println(s"${b.name}:\n ${b.value.pretty.render(100)}")
+                }
+
+                fail(s"Expected success, but got failure, logs=$logs")
     }
 
 }
